+++ src/main/java/com/eunsun/travel_mate/domain/PlanDetail.java
@@ -1,5 +1,6 @@
 package com.eunsun.travel_mate.domain;
 
+import com.eunsun.travel_mate.domain.tourInfo.TourInfo;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
