+++ src/main/java/com/eunsun/travel_mate/service/FavoriteService.java
@@ -1,6 +1,6 @@
 package com.eunsun.travel_mate.service;
 
-import com.eunsun.travel_mate.repository.FavoriteRepository;
+import com.eunsun.travel_mate.repository.jpa.FavoriteRepository;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
