+++ src/main/java/com/eunsun/travel_mate/service/PlanService.java
@@ -1,6 +1,6 @@
 package com.eunsun.travel_mate.service;
 
-import com.eunsun.travel_mate.repository.PlanRepository;
+import com.eunsun.travel_mate.repository.jpa.PlanRepository;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
