+++ src/main/java/com/eunsun/travel_mate/service/TokenBlacklistService.java
@@ -2,7 +2,7 @@ package com.eunsun.travel_mate.service;
 
 import com.eunsun.travel_mate.domain.TokenBlacklist;
 import com.eunsun.travel_mate.dto.request.TokenBlacklistRequestDto;
-import com.eunsun.travel_mate.repository.TokenBlacklistRepository;
+import com.eunsun.travel_mate.repository.jpa.TokenBlacklistRepository;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
